+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 -   Add to be planned events tab on calendar page
 -   Add employee selector to update intervention form
 -   Add employee filters on calendar component for admin
+-   Add event modal displayed on event card click on calendar component
 
 ## Fixes